Programmable Swings
Enabled swings to be set via MQTT and saved as a settings file.

Init.lua file was updated with a timer because for some unknown reason
that will probably bite me later, without it we would get an "out of
memory" error. The full code would run fine manually. It was just during
bootup that init.lua would give me the out of mem on dofile("main.lua").
Even running dofile("init.lua") worked when run manually, it was only on
bootup. The two second timer seems to have solved the problem.