New IO Logic for Rev 3 Hardware
Updated the bit logic for control of the relays. There were also some
typos in this version leading me to believe it was not the latest
version. But, I couldn't find a newer version so this is where we're at.
There's some debug code still in here, also, that can be removed.