Added check for running before setting mode
Noticed a logic neglected to check for whether the system was on or off
before setting the next state. This meant we were repeatedly calling the
"start" and "stop" routines even though the system was already started
or stopped. This didn't seem to cause problems, but if we're going to
change the behavior of the display we'll need this to be consistent.

I'm going to end the cleanup-0.01 branch and start a whole new branch
focused on lowering the code footprint and cleaning up the state
machine. I've been experiencing intermittent "out of memory" issues that
I can't explain. And, it has been so long I've lost track of the logic
in the state machine so might as well get that back in the box.