New LCD module, lotsa debug - but fits in memory
Another iteration of the rebase. Added lcd module separate in an effort
to minimize memory usage. It only refreshes portions of the screen live
instead of all of the screen from globals. The downfall is that the
values onscreen aren't necessarily saved anywhere. This is going to
cause a problem we we implement a user menu. We'll tackle that when we
get there.

Left the "print(node.heap())" statements in to monitor memory usage.